pyproject.toml
[project]
name = "linux-productivity-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyinstaller>=6.16.0",
    "pyqt6>=6.9.1",
    "pytesseract>=0.3.13",
]
